ADD: Moved group target encoder tests to own file and added test for BROS target encoder
